Refactor RELEASE
