--- .idea/artifacts/web_Web_exploded.xml
@@ -1,13 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="exploded-war" name="web:Web exploded">
-    <output-path>$PROJECT_DIR$/out/artifacts/web_Web_exploded</output-path>
-    <root id="root">
-      <element id="javaee-facet-resources" facet="web/web/Web" />
-      <element id="directory" name="WEB-INF">
-        <element id="directory" name="classes">
-          <element id="module-output" name="web" />
-        </element>
-      </element>
-    </root>
-  </artifact>
-</component>
