+++ service/pom.xml
@@ -18,7 +18,14 @@
     </properties>
 
     <dependencies>
-
+        <dependency>
+            <groupId>jakarta.servlet.jsp.jstl</groupId>
+            <artifactId>jakarta.servlet.jsp.jstl-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
         <dependency>
             <groupId>com.arinc</groupId>
             <artifactId>database</artifactId>
